write api tests
