+++ jhub_apps/tests/test_api.py
@@ -1,21 +1,36 @@
-import os
-from unittest.mock import patch, MagicMock
-from jhub_apps.service.routes import HubClient
+from unittest.mock import patch
+
+from jhub_apps.hub_client.hub_client import HubClient
+from jhub_apps.tests.constants import MOCK_USER
 
 
-PATCH_ENV = {"JUPYTERHUB_API_TOKEN": "secret-token"}
+@patch.object(HubClient, "create_server")
+def test_api(create_server, client):
+    from jhub_apps.service.models import UserOptions
+    create_server_response = {
+        "user": "aktech"
+    }
+    create_server.return_value = create_server_response
+    user_options = {
+        "jhub_app": True,
+        "display_name": "Panel App",
+        "description": "description",
+        "thumbnail": "",
+        "filepath": "",
+        "framework": "panel",
+        "custom_command": "",
+        "conda_env": "",
+        "profile": "",
+    }
+    body = {
+        "servername": "panel-app",
+        "user_options": user_options
+    }
+    response = client.post("/server/", json=body)
+    create_server.assert_called_once_with(
+        username=MOCK_USER.name,
+        servername='panel-app',
+        user_options=UserOptions(**user_options)
+    )
+    assert response.json() == create_server_response
 
-
-@patch.dict(os.environ, PATCH_ENV, clear=True)
-@patch.object(HubClient, "get_user")
-@patch("jhub_apps.service.routes.get_hub_oauth")
-def test_api_server(get_hub_oauth, get_user, client):
-    user_for_token = MagicMock()
-    auth = MagicMock()
-    user_for_token.return_value = {"name": "username"}
-    get_hub_oauth.return_value = auth
-    auth.user_for_token = user_for_token
-    returned_user_from_get_user = {"servers": [{"name": "server-name"}]}
-    get_user.return_value = returned_user_from_get_user
-    response = client.get("/server")
-    assert response.json == returned_user_from_get_user
